+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-01-17
+### Added
+- Added the ability to control cookiejar saving. Autosave can be disabled
+  and the session cookies can now be saved at any time, not just on close.
+
 ## [0.4.0] - 2022-09-12
 ### Added
 - Added the ability to do initial authentication with username/password